feat(frontend): change imports and exports page to data
